Give up on setting a "reasonable" TIMESTAMP_IMAGE
We want devs to be able to easily build this image locally and set a
reasonable TIMESTAMP_TAG. Unfortunately, we can't really know this
value a priori and would have to make lots of assumptions about what
the dev will tag the image with, what project they will be using, etc
etc.

Instead, just set this to "UNKNOWN" so that it jumps out at us if
somehow we get a locally-built image running in the OSPool.